First changes
First update for assignment code, the code was created before and as no entry was made then i’m updating it now. I have made the main page of the tkinter gui and have started the function for a subsequent page. The main page has placeholders for where each component should be and I have drawn out a simple diagram for what I plan to do. The only challenges when completing this were customising the entries, and buttons and placing/resizing the images which was later resolved.

I have also created a terminal based code (no gui so no tkinter), this code was to test if what I plan to do for my quiz would work, i have written it out and tested it, and it has worked. There weren't any challenges in this part of the production, however I did use copilot to write out my lists, this was only because of the time it would take to write it out on my own. An example is shown below.