Changing to custom tkinter, and changing the quiz style.
The next few changes are me changing the quiz's style which was also previously changed in canva, the code has also been changed into custom tkinter format. This was so I am able to created curved entry boxes for my options. I had changed this already, and copied it over to another doc, this is why i copied everything over from the doc, this was so I could keep the custom tkinter file while also saving the normal tkinter one.