Completion of quiz page
I have completed the quiz page, and have begun the final score page. I have added a progress bar, and bound some actions to the enter key so instead of pressing a button the user can also press enter to make things quicker. I have had some problems with the images, where some images are warped in scale, this is only with pages other than the front page of my app, the image of the front page also matches in perfectly with the scale on every other toplevel, this tells me that the problem is in canva, and not in the actual code. I have decided that if i have enough time left over i will change each image pillow format so I can resize them to the window screen, however they will not work on mac. The main problem I faced was with the infinite looping of the quiz, the quiz was meant to count off the iterations then when the iterations matched the question num, they were meant to configure a new button to move to a new page. There was an error in the syntax and this wasn't working properly, however I managed to get it fixed today.