Completion of the draft 1 code
The finish page has been completed. And excluding error messages, the code has also been completed. There was only one difficulty with this part of the code and that was the creation of an animated flowing progress bar to show the users proficiency. I scrapped a three layered version of this as it was far too complicated and required more advanced coding knowledge than we had learnt. I changed it to a simple one line progress bar, which changes colour as it progresses showing the user's level of proficiency and then displaying their quantitative result in the space above the bar. To complete the rest of the code I will spend time on cleaning up the code, and adding error messages. If there is time I will also optimize certain parts of the code.