Error messages and centering
I have centered everything in my code to frames so it will operate even when different screen sizes are used, as long as they are larger than 1920x1080. I have also completed all the error checks for the options page and decided to add a about page. I am however having issues with the error checking in the submit page. I plan it so if the entry for the quiz page is empty and the submit button is pressed, then there is a message saying to fill it in. if the submit button is pressed once, any other time it is pressed it will be said, please press next, and if next is pressed when the submit button hasn't been pressed, it will say press the submit button, the enter keys will just work normally if you just press enter.