Finishing the code
I have created all error messages and have completed my code, there are only some small errors and one large error. The large error is when the error message (an exterior module) temporarily removes the page if it is a toplevel. This has no effect on the functionality of the code, it just doesn't look right. I will remove this bit. I have also rearranged everything in the directory to make it more organised and have changed the relative path of the files to fit these changes.