Small edits
I made some small edits to the code, like formatting the code and deleting non essential things, like dev tests. I also added a contingency to hopefully fix one of my errors, it is very difficult to test, as it is rare and happens to a suddenly slow, or very quick input of details which in some cases, does not create the symbol/name prompt for the quiz page. I have fixed this by checking if that symbol exists or not and if it doesn't reloading the page.