final edits
For the code I added restrictions to the options page entries and then made some small changes. There was one error that was difficult to sell, which was the fact that the text course wouldn't focus on the quiz page entry in the first iteration, but this was later fixed.